refactor(bootstrap): group subsystem assemblies under subsystem/
Phase 4.2 of the uc-bootstrap redesign. Move the four sub-assembly modules
into subsystem/: sync_engine and analytics keep their names, while
file_transfer_lifecycle -> file_transfer and background_tasks -> blob_tasks
get clearer names. Contract re-exports (compose_event_context,
spawn_blob_processing_tasks/BlobProcessingPorts, FileTransferLifecycle,
SyncEngineAssembly) now source from subsystem::; internal crate:: paths
updated. No behavior change.
