+++ crates/uc-bootstrap/src/builders.rs
@@ -24,7 +24,7 @@ use uc_core::config::AppConfig;
 use crate::assembly::{
     get_storage_paths, wire_dependencies, BackgroundRuntimeDeps, SharedRuntimeDeps, SyncEngineDeps,
 };
-use crate::sync_engine::{build_sync_engine_assembly, SyncEngineAssembly};
+use crate::subsystem::sync_engine::{build_sync_engine_assembly, SyncEngineAssembly};
 
 /// Shared core wiring for the CLI composition-root entry.
 /// Initializes tracing, resolves config, wires dependencies, and registers the
@@ -73,7 +73,7 @@ async fn build_core(
     // 旧 CLI 进程内路径不是临时 daemon residency → 永不抑制设备级 presence 事件
     // （ADR-008 D20），保持 pre-P5 行为；P5-1/P5-2 会整体退役这条路径。
     if let Err(err) =
-        crate::analytics::compose_event_context(&wired.deps, &storage_paths, false).await
+        crate::subsystem::analytics::compose_event_context(&wired.deps, &storage_paths, false).await
     {
         tracing::warn!(
             error = %err,
@@ -86,7 +86,7 @@ async fn build_core(
 
 /// CLI composition-root entry returning the full
 /// [`crate::assembly::WiredDependencies`] so the caller can hand it to
-/// [`crate::sync_engine::build_sync_engine_assembly`]. It preserves access to
+/// [`crate::subsystem::sync_engine::build_sync_engine_assembly`]. It preserves access to
 /// `trusted_peer_repo` and other ports the `SpaceSetupFacade` needs (pairing /
 /// roster / send / watch / blob 等需要 iroh 网络栈的 CLI 命令走这条路径)。
 pub(crate) async fn build_cli_wiring_context(
