+++ crates/uc-bootstrap/src/lib.rs
@@ -5,16 +5,13 @@
 //! All entry points (GUI, CLI, daemon) depend on uc-bootstrap
 //! for dependency wiring and initialization.
 
-pub mod analytics;
 pub mod assembly;
-pub mod background_tasks;
 pub mod builders;
-pub mod file_transfer_lifecycle;
 mod network_policy;
 pub mod non_gui_runtime;
 pub mod observability;
 pub mod startup;
-pub mod sync_engine;
+pub mod subsystem;
 
 // The top-level re-exports below ARE the crate's external contract: the symbols
 // daemon (apps/daemon) and the CLI dev-tools feature (apps/cli) consume. Keep
@@ -23,19 +20,19 @@ pub mod sync_engine;
 
 // Slice 6 / Issue #549 — composition-root analytics 装配入口。
 // `compose_event_context` 在 `wire_dependencies` 之后由各进程入口调用一次。
-pub use analytics::compose_event_context;
+pub use subsystem::analytics::compose_event_context;
 
 pub use assembly::{
     get_storage_paths, wire_dependencies, BackgroundRuntimeDeps, SystemClipboardWiring,
     WiredDependencies, WiringError, WiringResult,
 };
-pub use background_tasks::{spawn_blob_processing_tasks, BlobProcessingPorts};
 pub use builders::{build_daemon_lifecycle, DaemonLifecycle};
-pub use file_transfer_lifecycle::FileTransferLifecycle;
 pub use non_gui_runtime::{
     build_app_facade_from_deps, build_cli_app_runtime, build_mobile_sync_facade,
     resolve_clipboard_integration_mode, AppFacadeAssemblyOptions, CliAppRuntime,
     ClipboardRestoreAssembly,
 };
 pub use observability::tracing::{init_tracing_subscriber, install_panic_logging_hook};
-pub use sync_engine::SyncEngineAssembly;
+pub use subsystem::blob_tasks::{spawn_blob_processing_tasks, BlobProcessingPorts};
+pub use subsystem::file_transfer::FileTransferLifecycle;
+pub use subsystem::sync_engine::SyncEngineAssembly;
